pass yield

+++ kernel/include/defs.h
@@ -136,7 +136,7 @@ int             either_copyin(void *dst, int user_src, uint64 src, uint64 len);
 void            procdump(void);
 void            test_proc_init(int);
 int             clone(int flags, uint64 stack, int ptid, unit64 tls, int ctid);
-int             wait4(int pid, uint64 addr, int options);
+uint64             wait4(int pid, uint64 addr, int options);
 
 // swtch.S
 void            swtch(struct context*, struct context*);
