+++ kernel/include/proc.h
@@ -98,6 +98,6 @@ void            procdump(void);
 uint64          procnum(void);
 void            test_proc_init(int);
 int             clone(int flags, uint64 stack, int ptid, uint64 tls, int ctid);
-int             wait4(int pid, uint64 addr, int options);
+uint64          wait4(int pid, uint64 addr, int options);
 
 #endif
